pyproject.toml
[tool.poetry]
name = "vidtrim-streamlit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.34.0"
ffmpeg-python = "^0.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
